CELE-107 Code simplification
